Rename subwindow with timerwindow
