+++ src/reducers/timerwindow/index.js
