+++ src/stores/child.js
@@ -1,7 +1,7 @@
 import createLogger from 'redux-logger'
 import { createStore, applyMiddleware, compose } from 'redux'
 
-import childReducer from '../reducers/subwindow'
+import childReducer from '../reducers/timerwindow'
 
 const logger = createLogger()
 const finalCreateStore = compose(
